32bit, record.

+++ core/trace.cc
@@ -114,7 +114,7 @@ Symbol findSymbol(void* find) {
     const byte* elf = exe.data;
     const Ehdr& hdr = *(const Ehdr*)elf;
     ref<Shdr> sections = ref<Shdr>((const Shdr*)(elf+hdr.shoff),hdr.shnum);
-    const char* shstrtab = elf+sections[hdr.shstrndx].offset;
+    const char* shstrtab = elf+sections[(uint)hdr.shstrndx].offset;
     const char* strtab = 0; ref<Sym> symtab; BinaryData debug_line;
     for(const Shdr& s: sections)  {
         if(str(shstrtab+s.name)==".debug_line"_) new (&debug_line) BinaryData(ref<byte>(elf+s.offset,s.size));
